serving/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
scikit-learn
joblib
pydantic
numpy
pandas
